IWHSO work party managers in Connacht
These GeoJSON files are based on the data on IWHSO work party managers in Connacht collected by Susan Schreibman, Monika Barget and Pádraig MacCarron in 2018. Using government reports detailing the production of medical supplies in Ireland between 1915 and 1919, these women could be identified as local organisers in the collection of sphagnum moss in the west of Ireland. We collected information if the women were mentioned in both reports or just one of them, and where they resided/worked.