Daly Network Nodes / Correspondents
This file contains the nodes (that is senders and recipients of letters) in the Charlie Daly network. Additional information was partly acquired from the content of the letters, the archival finding aids provided by the Kerry Library, and genealogical resources. If you find any mistakes, please let us know. We are always eager to update our data.