Python script to read Letters 1916 JSON data dump
This Python script was written to access the transcriptions of letters in the Letters 1916-1923 collection in a JSON data dump of 2019. The transcriptions were nested in a series of dataframes, lists and dictionaries. A careful analysis of the actual data structure of the JSON files was necessary to extract selected information. Similar scripts were used to read letter titles or the authors' gender.