Python script for identifying letters by gender
This Python script was used to select all letters from a corpus of XML files that were written by women. The information on the authors' gender was stored in the Letters 1916-1923 metadata. Unfortunately, most XML files would not parse because transcriptions also stored in the files had been created by multiple volunteers and still required cleaning. This is why the Python package Beautiful Soup, written for XML as well as HTML, was used.